Simplify / document unsafe `extract_unary_type_arguments`
Summary:
This function needed a rename, but also needed to be clearly documented
as unsafe and only actually usable on types (in particular, a handful
of hardcoded, known types relevant to specific special case handling
in `typeCheck`) that are guaranteed to have only unary type parameters.

This matters because the implementation is already broken in general,
the behavior is undefined (I mean, it has some nonsense behavior I'm
sure but there's zero chance it is useful) if it were used on a
general-case generic type.

As part of this change, I eliminate all uses of the function directly
from `TypeCheck` - it should *only* be used in hardcoded functions
on known-to-be-okay types in `GlobalResolution.ml`. The function is
still exposed in the `.mli` file, but with `unsafe` in the name and
marked as for unit testing only.

In addition, the code is simplified because a number of special cases
aren't necessary in the first place once it's made explicit that this
function can only be used safely on hardcoded types in the first place.

Reviewed By: yangdanny97

Differential Revision: D61956518

fbshipit-source-id: 8020813706386f891db872d035bdf10f86773e04