Simplify `generic_parameters_as_variables` interface
Summary:
Simplify the function by making it clear that the `~default`, which
in principle allowed the injection of *anything*, is only actually ever
used to say "treat a non-generic class as if it were generic with
no parameters".

In addition, I was able to eliminate the ugly flag entirely from
`globalResolution.ml`s interace - there's *exactly* one use case
for the flag in `attributeResolution.ml`

All of this makes changes to the interface (I need to change the output
type in an upcoming diff) easier, as well as making it much more obvious
how the function is actually used than was the case when threading a
`default` value through.

The current interface still isn't optimal here I know, but it's a lot
better than what came before and more improvements would require a bigger
refactor that maybe isn't worth doing right now.

Reviewed By: yangdanny97

Differential Revision: D61994056

fbshipit-source-id: 0ca7d8debc529b9da18e442a67ee09a1aa0b81e8