Document a bug in frozen inheritance checks
Summary:
We are supposed to validate that there is no invalid frozen
inheritance chains, but the code is broken because we are relying on
a very dubious `Expression.show` to get the base class name, which can't
work for non-primitive base classes.

I ran into a refactoring problem in `typeCheck.ml` (which is related to
PEP 695 through a chain of blockers where the code architecture doesn't
make sense) that is actually a conformance bug.

Adding this test before I fix it, which in turn will unblock refactors
I need to make in order to eventually eliminate the assumption that variance
is obtained by extracting type arguments of `Generic`.

Reviewed By: yangdanny97

Differential Revision: D61994057

fbshipit-source-id: ef99d2d314e2c5d835e532934044c16a38409732