Separate computing vs checking base class types
Summary:
The existing code for how to get the base class types and perform
checks on them is a bit too complicated for me to easily see how to
extract variance checks from the rest of the logic.

A big part of the problem is that a bunch of the checks are running
concurrently with the extraction of the types, which isn't necessary
at all.

Before I attempt any bigger changes I just want to simplify this a bit.

Reviewed By: grievejia

Differential Revision: D61994062

fbshipit-source-id: 6c47c3025fa471acf99ba3f004ec992542bc290a