Decouple dataclass vs vairance checks, make helpers consistent
Summary:
This is the second of two diffs where I make the existing checks that rely on
base classes readable enough for me to more easily see the best way to extract
variance-related handling from the rest of the iteration.

In the process I am able to move a bunch of dataclass-specific helpers into
the dataclass inheritance check, which is really good for code-folding-oriented
skimming of the code (that's pretty important in the current type check module!).

This diff still has no functional change, it just makes each check independent
of the others and has them all use a similar type-level interface, which makes
it much easier to write a pipeline that's easy to skim.

(By "a pipeline that's easy to skim" I mean this block:
```
          empty_errors
          |> check_dataclass_inheritance base_types_and_locations
          |> check_variance_inheritance base_types_and_locations
          |> check_generic_protocols base_types_and_locations
          |> check_protocol_bases base_types_and_locations
```
which is quite a lot easier for me to follow than what we had before.)

Reviewed By: yangdanny97

Differential Revision: D62011608

fbshipit-source-id: 2a8ea0bbe2301d40f36815e0b1c43df5edf850ad