Create and update dataclasses for handling PTT type errors in IDE
Summary:
Lucian came up with a good idea around some extra metrics to include, and there are also a few that I think might be useful in the future:
* the type errors returned from Buck, so we can use for debugging later
* the error message from the Buck run, so that we can debug errors we're running into
* the Buck UUID, so that we can look into what takes some of the longer runs so long to complete
* if the GLTC type errors match PTT type errors, so we can determine the accuracy improvement between IDE and CLI/CI by switching to PTT.

This diff stack adds the functionality to support the above telemetry.

 ---
This diff adds two new dataclasses used to parse the type errors returned from PTT and updates the `PyreBuckTypeErrorMetadata` dataclass to hold this telemetry information.

Reviewed By: MaggieMoss

Differential Revision: D62045816

fbshipit-source-id: a8e88f3ae8957779614baf4f27a2437220bbd76e