Make variance checks obvious with renames
Summary:
I need to refactor the variance checks to not depend on parameters
being typed as `Type.Variable.t` (this is part of a stack to eliminate
the assumption that variance lives in that type at all, because trying
to model data that way is going to cause significant problems for
`parse_annotation` versus variance inference).

This diff is a trivial rename to make it clearer what's going on in
the first place; one set of "arguments" is really the generic parameters
of a base class and the other is the type arguments to the base class.
Both get filtered to just unary (TypeVar) arguments before the checks
run.

Reviewed By: yangdanny97

Differential Revision: D62016074

fbshipit-source-id: 85a47911473bc43025136ee0cd738dd547157c28