Make (some) variance checks easy to translate to GenericParameter.t
Summary:
This diff preps me to cut over the first callsite from using
`generic_parameters_as_variables` to using `generic_parameters`, by
eliminating the (it turns out, unnecessary) reliance on `Variable.all_unaries`;
we can just use a simple match in the final comparison.

As an added bonus, this change actually makes the code a bit simpler than with
the filtering (the match types are a little more complex, but the control flow
is much more straight-line without the filtering), and I suspect it's a tiny
bit faster due to fewer allocations.

(There are more places we'll have to update later that aren't in this diff,
cutting over `typeCheck.ml` entirely takes another 8 diffs or so... here I'm
just prepping portions of the current-class-vs-base-class check).

Reviewed By: yangdanny97

Differential Revision: D62016075

fbshipit-source-id: b31904d98e096d1af5347597b0b92ff98313435a