Add tests illustrating Pyre's protocol inheritance checks
Summary:
I regressed this on the conformance tests during a refactor, but got
no unit test failures. That's slightly concerning in itself; in addition
it appears that we are actually failing to handle the generic protocol
base case right now (lucky for us the conformance tests don't cover
that).

The next diff will refactor and also fix the bug in the process.

Reviewed By: yangdanny97

Differential Revision: D62131140

fbshipit-source-id: e2eac9a2bf6cd87ba40b2a3cb18cae00c6364643