Fix conformance bug: generic protocol is still protocol
Summary:
As part of the refactor in the previous commit, at one point I broke
our check that there should be no non-protocol bases for a protocol.

I the process of fixing that, I realized we had a preexisting bug:
if the protocol base is subscripted, the current code skips the check!

This is wrong, and I might as well fix it while I'm looking at this
code already.

Reviewed By: yangdanny97

Differential Revision: D62131143

fbshipit-source-id: bec9419d50562eb7d801d3794138629b731693cb