Prepare plumbing to cut type order to `generic_parameters`
Summary:
In this diff I'm continuing with a refactor stack to eliminate all uses
of `variance` coming from a `TypeVar.t`, which is a blocker on PEP 695
support because we need to be able to `parse_annotation` earlier in the
analysis than we can perform variance inference.

The remaining use cases of generics are inside of the guts of Pyre, in
the type order + constraints solver.

Since I want the cutover itself to be focused and easy to review, it is helpful
to separate the trivial plumbing changes to passing `generic_parameters`
as a callback; for now we can just map `GenericParameter.to_variable` in the
callsites.

Reviewed By: yangdanny97

Differential Revision: D62075940

fbshipit-source-id: 07b6f1b06b0ecd2096aa5bd5fe3b005e79b4a200