Move GenericParameter module down
Summary:
In order to cut over the core constraint solving and ordering
logic to using GenericParameter.t I'm going to need combinators
for zipping two arguments lists with a parameters list, which is
a fairly complex operation because of the structural pattern
matching aspects introduced by `TypeVarTuple` and `ParamSpec`.

I'll need some of the helpers defined in top-level (non-Record-namespaced
modules) of `type.ml` that aren't available right now because I
initially put `GenericParameter` close to the definition of `Type.t`.

This commit moves the module, so that the later changes will be
easier to read (when there's a move plus new logic added in the same
commit it can be hard to see in diff view). It's just a move, no
code changes whatsoever.

Reviewed By: yangdanny97

Differential Revision: D62075942

fbshipit-source-id: dd76e182a98676d513a3bff85ac568f7978c76be