Decouple AnalysisError's use of variance from TypeVar
Summary:
I intend to eliminate variance from the `TypeVar.t` datatype in
this stack after eliminating use cases, because we need to be able
to call `parse_annotation` before we perform variance inference when
supporting PEP 695.

For the most part these changes improve the structure of Pyre anyway,
because variance is a notion attached to a generic type *parameter*,
which is a related but separate concept from the *variable* that
is the name representing that paramter.

Here I'm modifying the invalid variance error to stop getting variance from a
`TypeVar.t` (both for the base class when there's an origin, and for this class
in all cases); instead we manually extract it from the `GenericParameter.t`.

An added benefit is that now I always actually show the variance. The previous
behavior of silently omitting "invariant" was not great - this error is
specifically about variance and many users don't even understand the concept
well, so imitting the keyword that they would need to google is not helpful.
I'm actually much happier with the errors now.

Reviewed By: grievejia

Differential Revision: D62029502

fbshipit-source-id: 48950af517b75ebe0bde58440dfd917d29078c56