Finish migrating typeCheck base-vs-this to variance-from-parameters
Summary:
This commit is the second half of decoupling the `typeCheck.ml` checking of
base class vs current class variance from the assumption that variance should
live directly in `TypeVar.t`
- an earlier diff got type parameter information for the *base* class from
  `GenericParameter.t` which was easier to do first because there's no extra
  lookup.
- this diff finsishes the cutover by adding a lookup step to go from the type
  *argument* to a base class (which is a type variable) to the type *parameter*
  of the current class.

Once again, all of this is necessary becasue we need `parse_annotation` to be
able to run well before variance inference in the processing pipeline. That means
we need to be getting variance information from somewhere principled, not trying
to pull them directly out of `Type.t` values wherever we have them.

-----------

This isn't quite the end of our `typeCheck.ml` changes, we still need to
cut over the parameter/return annotation variance checks.

Reviewed By: grievejia

Differential Revision: D62068858

fbshipit-source-id: ea52fe62d244ddafdc5a631d7e9f3e77dfff20aa