Clean up dead variance helpers in `Type.Variable`
Summary:
There are still two uses of variance-from-type-var we have to clean
up in the type order / constraint solver logic.

But all callsites of the `is_covariant` and `is_contravariant` helpers
were in `typeCheck.ml`, so we can clean these up.

Doing it now because it makes finding downstream logic easier if I eliminate
indirect usages (for example my initial refactor plan didn't account for some
parts of type check that needed fixing because they used these helpers and not
the field, so compiler-driven planning didn't work as well).

Reviewed By: yangdanny97

Differential Revision: D62069281

fbshipit-source-id: dc2a490910cf12ded5611d0b29ba97bae9409d3b