Add test case documenting that FPs due to unscoped variance checks
Summary:
I realized Pyre has a bug in the modeling of variance that will go
away when I make the changes needed to obtain variance from the
`GenericParameter.t` values that we get from the `generic_parametes`
query.

I wanted to get the test in before making the change so it's easy to see one of
the consequences of our incorrect scope-free data model that until the past few
weeks treated type variables, type parameters, and type arguments as
interchangeable, and how a better data model makes the incorrect analysis
impossible to even express.

Reviewed By: grievejia

Differential Revision: D62069187

fbshipit-source-id: 0a4342205053e75e5574c58de70411201478b06e