Rename `initial` uses of `legacy_parent` values
Summary:
The `leegacy_parent` value is, for the moment, only legacy
for some purposes, not others.

If you want define nesting information, you should use the
`NestingContext.t`, not `legacy_parent` which is where the term
`legacy` comes from.

But class tables are still keyed on (mangled) qualified names
as produced via qualification, which means `legacy_parent` is
currently very much the correct value to use when you want
the current class in a form where symbol table lookups will work;
until / unless we re-key all of those tables similar to what
we did in `FunctionDefinitionEnvironment`, this is not legacy
at all for such use cases.

But it's also a pretty bad name for such use cases. Sometime soonish
I'd love to rename it on the AST itself, but that's a distraction from
my current refactors; for now I am just seeing that naming it
`maybe_current_class_reference`, and naming various downstream values
(converted to Identifier.t, unpacked from the option) accordingly really
does make the code easier to skim... otherwise you'd have to understand
a lot of weird corners of Pyre before it's anywhere near obvious why
`legacy_parent` is getting used in this way (or at all, given that
it's marked as legacy!)

----------------

Reviewed By: grievejia

Differential Revision: D62068859

fbshipit-source-id: 89d1ca24bf460173fc8503e0a930b18acdacbfc7