Cut GlobalResolution `is_invariance_mismatch` to use `GenericParameter.t`
Summary:
This is very easy compared to the large-ish refactors that we've been
needing in order to cut over `typeCheck.ml`.

We were already mapping to extract the variance
after calling `all_unaries` so replacing with a `filter_map` and getting
the variance directly from `GenericParameter.t` is a drop-in replacement.

Reviewed By: grievejia

Differential Revision: D62069282

fbshipit-source-id: 4f65c10fdce39530e590cccfdafbe104ce092c2e