Cut return and parameter variance checks over to `GenericParameter.t`
Summary:
This is the payoff of the previous diff where I disabled variance checks
on generic functions (which matches Pyright's behavior and is permissible
under the typing spec): we can now

Note that there actually is a behavior change here: when looking at a generic
method of a generic class (where some type vars are bound on the class and some
only on the function), as of the previous diff we would still validate variance
for the method parameter. With this diff, that stops happening (becasue we are
powering variance checks off of the class type parameters, and the class
knows that there is no variance associated with a type var only bound on a
method).

This is a good change; the previous behavior was only an approximation to what
makes theoretical sense (see the removed TODO in the tests).

-----------------

Note: I realized as I worked on the code that the current variance checks are
extremely simplistic: we aren't checking polarity at all, we're only checking
*bare* type vars used as parameters or return types.

This is definitely not conformant, the correct variance checks would be
more involved.

It's unclear how important fixing this is in terms of our 2024 goals, given
that variance inference should lead to many fewer variance checks. But
it is definitely a gap.

When we do fix it, instead of making the exiting code more complex I'd at
least consider an alternative implementation of the checks where we just
infer variance and look for mismatches against explicit variance; that would
ensure (maybe at some perf cost?) that we only have to write the polarity logic
once instead of having two copies of it.

Reviewed By: grievejia

Differential Revision: D62069188

fbshipit-source-id: 9b4bf0c123d8d760cfbe12a0b21a713e3e43c1d3