Parse PTT type errors in IDE
Summary:
Lucian came up with a good idea around some extra metrics to include, and there are also a few that I think might be useful in the future:
* the type errors returned from Buck, so we can use for debugging later
* the error message from the Buck run, so that we can debug errors we're running into
* the Buck UUID, so that we can look into what takes some of the longer runs so long to complete
* if the GLTC type errors match PTT type errors, so we can determine the accuracy improvement between IDE and CLI/CI by switching to PTT.

This diff stack adds the functionality to support the above telemetry.

 ---
This diff calls the PTT error parsing logic at the end of the PTT type check function.

Reviewed By: grievejia

Differential Revision: D62045819

fbshipit-source-id: 6458be04d4909e6c31a597067a07ef53e1728087