Add telemetry around PTT type errors and GLTC accuracy
Summary:
Lucian came up with a good idea around some extra metrics to include, and there are also a few that I think might be useful in the future:
* the type errors returned from Buck, so we can use for debugging later
* the error message from the Buck run, so that we can debug errors we're running into
* the Buck UUID, so that we can look into what takes some of the longer runs so long to complete
* if the GLTC type errors match PTT type errors, so we can determine the accuracy improvement between IDE and CLI/CI by switching to PTT.

This diff stack adds the functionality to support the above telemetry.

 ---
This diff adds the type errors found in the previous diff to the telemetry we return to VSCode.

Reviewed By: grievejia

Differential Revision: D62045813

fbshipit-source-id: 0b73d39dcaf8268f15b61d8ac13f7523a110cf66