Return early if the PTT Buck call errors out
Summary:
Lucian came up with a good idea around some extra metrics to include, and there are also a few that I think might be useful in the future:
* the type errors returned from Buck, so we can use for debugging later
* the error message from the Buck run, so that we can debug errors we're running into
* the Buck UUID, so that we can look into what takes some of the longer runs so long to complete
* if the GLTC type errors match PTT type errors, so we can determine the accuracy improvement between IDE and CLI/CI by switching to PTT.

This diff stack adds the functionality to support the above telemetry.

 ---
This diff returns early if the Buck call fails for a reason other than being preempted.

Reviewed By: grievejia

Differential Revision: D62047055

fbshipit-source-id: c3a915ba03bbac8ba39490476ee3b86b80dca26f