Create new zip combinator for GenericParameter
Summary:
This commit is part of a stack to eliminate the dependence of
our code on variance obtained from `TypeVar.t` values; we need
the variance to stop being treated as part of the variable itself
(which is just a name) and instead treat it as part of a generic
parameter itself in order for variance inference to run after
`parse_annotation`.

This commit makes a `GenericParameter.t` version of the legacy
combinator `zip_variables_with_two_argument_lists`.

This is quite a lot of code because zipping is a complex operation
for Python; unlike most languages where we could just zip, we are
actually doing a fairly complicated manual pattern match in which
both the parameters *and* the arguments can potentially contain
- unpackable expressions / type var tuple slots, which make the
  zip not even close to 1:1
- callable parameter / param spec slots, which are distinct from
  other kinds of variables

The complexity here is pretty similar to matching function calls and
signatures; there too a call that might or might not contain a `*args` argument
with a signature that might or might not have a `*args` parameter. This
is just the type-level equivalent of the same problem.

The new code is considerably more type-safe than before, the entire
pattern matching space of various successful and unsuccessful zips
is modeled to the best of my ability.

NOTE: Operating on lists makes the pattern-matching code trickier; there are
invariants on both the argument and parameter lists that they should not
contain multiple variadic forms (Unpackable / GpTypeVarTuple) but the type
system can't see these (and honestly I can't even tell whether these invariants
are actually enforced upstream of this logic). Having optimal datatypes for
this kind of complex argument/parameter pattern match is probably a good thing
to explore if we rewrite the `Type.t` datatype; similar issues happen in
callable parameters vs args for similar reasons (*args plays the same role there).

Reviewed By: migeed-z

Differential Revision: D62075943

fbshipit-source-id: 43d9e26366123681aba73bd39d9874ec885a247b