Delete variance from TypeVar.t
Summary:
The entire stack has been working up to this point: variance is not
in fact a property of a type variable, but rather of the type parameter
on the generic that binds it.

Pyre modeled this incorrectly which is understandable because until
PEP 695 the Python *syntax* modeled it incorrectly, but with PEP 695
and variance inference we had to fix this problem, because we
have to translate expressions to types *well* in advance of running
variance inference, which means trying to get variance from the
type vars themselves could not possibly work.

The refactor was massive, but along the way we:
- Improved Pyre's data model quite a bit by making a distinct datatype
  of a `GenericParameter.t` to model the notion of type paramters (which
  are not the same as variables or type arguments; Pyre often conflated
  these concepts prior to recent refactors)
- Fixed a few actual bugs in cases where the existing code couldn't
  reasonably be migrated to the new data model.

At this point the plumbing part of adding PEP 695 support to generic
classes and adding variance should be relatively easy, we just have
to extend the way `generic_parameters` is computed (ideally by
also splitting `GenericParameter.t` into two datatypes, one whose
variance is not-yet inferred - let's not repeat the mistake of
failing to put ocaml type safety into this kind of thing!)

Reviewed By: migeed-z

Differential Revision: D62076557

fbshipit-source-id: 526cbe05c7d365a845bb5ebe71a6fec1c25c0533