Fix (mostly) a conceptual gap in TypeAliasEnvironment
Summary:
**History / existing state:**

A couple of months ago we realized that TypeAliasEnvironment
frequently stored type variables as `TypeAlias` variants instead
of `VariableDeclaration` variants. We realized this would make
working out the changes for PEP 695 and scoped type variables nearly
impossible, so we fixed the problem.

But until this commit there was a case that still leads to a `TypeAlias`
containing a bare `Variable`: we currently allow users to import type variables
(I think this is unfortunate, and what is really bad is that when combined with
Pyre's "implicit generic" behavior which is a spec violation but some code
relies on is impossible to tell by looking at the code whether a class is
generic at all, you have to have global knowledge of all imports).

Even worse, the name of the type variable was actually wrong: a legacy
hook inside of `unchecked_resolve` was overwriting the name with the
fully-qualified name of the import alias.

This is both a major BE issue because the Pyre codebase doesn't behave the way
you might reasonably expect on import alias chains for type vars, and it also
blocks us from migrating `ClassHierarchyEnvironment` away from relying on
variance coming directly out of a `Type.t` because there are examples of
generic classes using chained imports of type variables in real code.

**This fix**

The TypeAliasEnvironment logic is too convoluted for me to figure out an optimal
fix right now; my instinct is actually that a chunk of the logic needs to be
migrated down into UnannotatedGlobalEnvironment and then we need *much* clearer
semantics than we currently have for the graph traversals; we have to revisit
this anyway if we want to address conformance failures.

Because of this, I'm avoiding trying to rewrite the code; instead I am just
making two targeted changes:
- stop rewriting the name of type variables in `unchecked_resolve`; this
  was here to provide best-effort support when users declare a type variable
  with the wrong name (`T = TypeVar("S")`) and hasn't been neeed since we
  introduced `Type.Variable.Declaration`, which handles that problem.
- at the very end of the `get_aliased_type_for` function, if we discover that
  we have an alias for a bare type variable we intercept the alias and
  convert it into a `RawAlias.VariableDeclaration` by looking it up directly
  in the aliases map; we know that we'll find a declaration because that's
  the only way to have gotten a `Variable` in the first place.

**Here be Dragons: there are still bugs somewhere in this code**

NOTE: There is some edge case I have not managed to reproduce in tests that breaks our
attempt to get a declaration. On a large production codebase there are exactly 3 type
variables - all of them coming from thrift-related code - hit the None case where we have a
bare `Type.Variable` in an alias but we have no corresponding declaration. I don't know
how this can happen yet, my attempts to find a repro have thus far failed.
The workaround I am using for the moment is to log when it does occur
and fall back.

It turns out that the fallback is doing no harm at the moment, even though it could.
This makes me think we should land and maybe fix-forward; this stack needs to land pretty
quickly both to avoid rebase hell and unblock PEP 695 work, and given that the remaining
bug seems to have no impact on existing code we can afford to root cause it after landing
rather than before.

Reviewed By: migeed-z

Differential Revision: D62083215

fbshipit-source-id: 50334dbbf5a8ebd0a0d6881ead179c01fecce9bc