Cut ClassHierarchyEnvironment away from TypeVar.variance
Summary:
This diff is part of a stack to remove variance from TypeVar.t,
which is a blocker for PEP 695 syntax support because in order
to do variance inference we need `parse_annotation` to not depend
on the results of variance inference (which has to run later in
analysis).

This is finally the very last "true" legacy use case of TypeVar.t variance in
production code; remaining use cases are all either in unit tests (which are
brittle across this sort of change) or in places where we explicitly made up a
fake variance to get started on PEP 695 work before this refactor was complete.

Instead of obtaining the variance from the `TypeVar.t`, we now fetch the
variable declaration from the environment and use the variance there. This is
conceptually much more correct: legacy-style type variables do not have
variance, their *declarations* specify variance that should be attached to the
relevant generic parameters of type constructors, but are not properties of the
variables themselves (a variable itself is just a name).

---------

Reviewed By: migeed-z

Differential Revision: D62075944

fbshipit-source-id: b67d92637ec4c0b396f80305a3315777a986f59c