Migrate Test.connect and unit tests that use it
Summary:
This is one of the last diffs in a stack to remove variance from
`TypeVar.t`, which we are doing because it blocks PEP 695 variance
inverence (we cannot be assuming variance in `parse_annotation`, because
the variance has to be computed later).

The `Test.connect` helper function exists only for a relatively
small set of unit tests. It is brittle in that it assumes
variance is a part of type variables inherently, so it has to
be migrated across the removal of variance from `TypeVar.t`.

I extended the `connect` helper into a `connect_with_variance` that allows
passing a variance with every argument; most test cases can use the default
`connect` which will automatically assume all arguments are invariant.

I also included a sanity check that the variance on a type variable and
the variance explicitly passed always match; I'll delete that check shortly
when I remove variance from `TypeVar.t` entirely but it's nice to have
the proof that this migration worked.

Reviewed By: migeed-z

Differential Revision: D62076559

fbshipit-source-id: 232645da03c787686a06a27542ab4b11a8d02130