fix callable subtyping bug in simulated signature select
Summary:
A function with `**kwargs: T` should be callable as a function with named params if all the named params that would get matched with the kwargs are compatible with `T`.

Previously, if a kwargs on the left doesn't get matched to another kwargs we would just move on, and the leftover named params on the other side would cause the constraint solving to fail.

This problem is documented in the attached task, which is partially fixed by this diff.

Reviewed By: grievejia

Differential Revision: D62200534

fbshipit-source-id: 0d28db2f07a773087fb9e9e9b6c467c8a0bbdca3