Support class definitions
Summary:
In the previous diff, we collected type params from the AST and into class definitions. Following this, we:

1- Add the type parameters in the classHierchyEnvironment by recording it as a generic parameter, using the new interface which Steven created.

2- In attributeResolution, we get the parent of our current defintion (but use the legacy parent due to qualification) and we retrieve the parameter list and add it to the scope

Note:  typechecking isn't working well with PEP695 syntax and I'm not sure why that is yet. AttributeResolution's global_annotation behaves as expected, so I wonder if this is more complicated and has to do with constraint solving. I explained this in the comments on this diff, but this is why I had to annotate my variable in the unit test. Once we propagate the RHS type to the LHS, we should be able to get rid of the annotation.

Reviewed By: stroxler

Differential Revision: D62162855

fbshipit-source-id: 459075a5cc836329ed99d350e4049331b2d927fa