handle list/set/sorted in weakenMutableLiterals
Summary:
the header of weakenMutableLiterals mentions that it should be done for constructors

> ... that doesn't hold for literals (or constructors, in general)

Right now the analysis only handles literals and comprehensions, but I think we can afford to weaken a few more things here to improve devx.

The list & set constructors + the sorted stdlib function all create new containers, so they should be safe to weaken.

This diff adds weakening support for them, along with a simple `sorted` stub in test.ml

Reviewed By: samwgoldman

Differential Revision: D61854007

fbshipit-source-id: c0654f0676635e48503c2e7b67c460463251342e