Include the query in failed queryTest.ml assertions
Summary:
Sometimes I'm amazed we ever managed to build thousands of lines worth of unit
tests in which there is nothing in the error itself (in this case dozens of
lines of json that don't even look like the values hardcoded into OCAML)
mapping to the test case, and the stack trace from OUNIT doesn't point to the
test failure, and there is no error message.

I don't understand how to effectively debug anything in that world. These tests
are Lwt-ified so I don't think the usual test-splitting technique will work, but
what's very easy to do is include the query (which unlike the results is searchable,
it's a string in the unit test source code) in the error message.

Reviewed By: grievejia

Differential Revision: D62385226

fbshipit-source-id: 5451bfa2915c4a8403b5d749a2c26fccd7272813