Improve assertion names in queryTest.ml
Summary:
I found `assert_query_response_with_local_root` to be not
a very helpful name, and `assert_type_query_response_with_local_root`
was actually actively misleading, it was used for queries other than
the `types(...)` query.

Renaming them to emphasize the actual difference - one takes the response
as json, the other takes it as an ocaml value with a type.

Reviewed By: grievejia

Differential Revision: D62384771

fbshipit-source-id: 0a5cf2fd7c2e9b8ef3b58df5bb02480d8f884c1d