Try running PTT in different isolation directory non-preemptible
Summary:
Going to test not preempting at all to see if there's a difference in running times.

Jia proposed the possibility that preempting might throw away all partially generated state, even if Buck is close to finishing. So maybe not preempting could give us a speed boost in some cases.

This diff stops us from preempting so that we can measure the difference.

Reviewed By: MaggieMoss

Differential Revision: D62461741

fbshipit-source-id: b2be71c67fd4955d874db31431730169d35fd500