Use the `Context.record_expression_type` hook in `forward_expression`
Summary:
The previous diff added a new `Context` callback we can use to properly record expression
types that we traverse in `typeCheck.ml`, which should allow us to rewrite the `types(...)`
query to just retrieve this data.

We expect that this change ought to fix all gaps in the `types(...)` query on
code that gets traversed in type checking (which is not everything; for example
lots of signature ast nodes aren't traversed by type check and will still
require special handling).

In the next diff I'll modify the context for `compute_local_annotations` to actually
fill out a table of expression / type pairs.

Reviewed By: rchen152

Differential Revision: D62449019

fbshipit-source-id: 36ee6ffa8cce15bbd5b53a877ed24c8f2ba9d027