Add expression/type saving callback to Context.
Summary:
I was looking at our options to support a partner team that needs
better results from the `types(...)` query and concluded that
we probably need to completely rewrite the expression visitor, which
can't handle complex expressions because the `Resolution.t` value can
actually change as we traverse expressions, which means that we
need to do that traversal in the context of the actual type inference
recursion.

This means that one way or another, all options involve some form of
dependency injection. I considered a few options, but then I realized
it's kind of obvious: the `Context` module is already where we are injecting
this kind of thing, for example the `Callgraph.Builder`. Given that this
is the existing pattern, we ought to just stay consistent.

This diff just wires up a slot for types to be saved, I'll use it later.

Reviewed By: rchen152

Differential Revision: D62449020

fbshipit-source-id: 7c8af3d25d0407ecbd04c8091ba7221b56fefad7