Understand that type variables are bound in preprocessing
Summary:
To unblock typechecking for PEP695, we have to allow preprocessing to understand that type variables are bound variables.

We do so for top level function defintions and type statements. We will repeat the same thing for classes in the next diff. (It doesn't work to simply do it in the class branch of this function, so this is pending investingation).

This changes our type errors to ones of the form
"Name `T` is used but not defined in the current scope." to "Annotation `T` is not defined as a type.", which we can then fix in our typechecker.

For type statements, I think we should also remove the name itself from the access set. But it doesn't really work to do it where we're currently handling type statements (the type alias branch). So this is pending investigation.

Reviewed By: grievejia

Differential Revision: D62469695

fbshipit-source-id: 4d6f92f63f52917e4898683412701e4976baf07c