Short-circuit potentially expensive check
Summary:
Calling `type_contains_untracked_name` checks if a class exists for every name
in an annotation. Checking if a class exists reads the entire class summary,
although it ignores the result and just returns `true` if the read returns
`Some _`.

I noticed this while reading the code investigating a performance regression.
This was not the root cause, but this change seems worthwhile regardless.

Reviewed By: rchen152

Differential Revision: D62510857

fbshipit-source-id: 6a0d77a96513989bda5b0e220de98a167e44aabd